requirements.txt
dj-database-url==0.3.0
Django==1.11.6
django-filter==1.0.4
djangorestframework==3.7.0
Markdown==2.6.9
python-decouple==2.3
pytz==2017.2
requests==2.7.0
Unipath==1.1
